+++ src/main/java/christmas/domain/discount/Discount.java
@@ -4,7 +4,7 @@ import christmas.domain.Orders;
 import christmas.domain.ReservationDate;
 
 public interface Discount {
-    long calculateDiscount(Orders orders, ReservationDate date);
+    int calculateDiscount(Orders orders, ReservationDate date);
 
     DiscountType getType();
 }
