+++ src/main/java/christmas/domain/discount/DiscountDetail.java
@@ -1,5 +1,5 @@
 package christmas.domain.discount;
 
-public record DiscountDetail(DiscountType type, long amount) {
+public record DiscountDetail(DiscountType type, int amount) {
 
 }
