+++ src/main/java/christmas/domain/discount/WeekDayDiscount.java
@@ -10,12 +10,12 @@ import java.time.DayOfWeek;
 public class WeekDayDiscount implements Discount {
 
     @Override
-    public long calculateDiscount(Orders orders, ReservationDate date) {
+    public int calculateDiscount(Orders orders, ReservationDate date) {
         DayOfWeek dayOfWeek = date.getDayOfWeek();
         if (isWeekday(dayOfWeek)) {
             return calculateDiscountAmount(orders);
         }
-        return 0L;
+        return 0;
     }
 
     @Override
@@ -26,8 +26,8 @@ public class WeekDayDiscount implements Discount {
     private boolean isWeekday(DayOfWeek dayOfWeek) {
         return !dayOfWeek.equals(DayOfWeek.FRIDAY) && !dayOfWeek.equals(DayOfWeek.SATURDAY);
     }
-    private long calculateDiscountAmount(Orders orders) {
-        return orders.getOrderItems().stream()
+    private int calculateDiscountAmount(Orders orders) {
+        return (int)orders.getOrderItems().stream()
                 .filter(orderItem -> "Dessert".equals(orderItem.getMenuItem().getCategory()))
                 .mapToLong(OrderItem::getQuantity)
                 .sum() * WEEKDAY_DESSERT_DISCOUNT.getDiscountAmount();
